Major reorganization of files. Writing code documentation in progress Moving explosion and projectile type definitions to data files (loss of ion EMP logic to be replaced)

+++ src/Primrose/Primitives/Interfaces.cs
@@ -3,10 +3,18 @@
   /// <summary>
   /// Implements an ID object
   /// </summary>
-  public interface IIdentity { int ID { get; } }
+  public interface IIdentity
+  {
+    /// <summary>The instance ID</summary>
+    short ID { get; }
+  }
 
   /// <summary>
   /// Implements a named object
   /// </summary>
-  public interface INamedObject { string Name { get; } }
+  public interface INamedObject
+  {
+    /// <summary>The instance name</summary>
+    string Name { get; }
+  }
 }
